Create image utils for base64 and PIL transformation, move processing of outputs and inputs to model utils with lazy import
